add qaqc.py in core
