+++ stglib/rsk/cdf2nc.py
@@ -2,7 +2,7 @@ import numpy as np
 import xarray as xr
 
 from ..core import utils
-from .. import exo
+from ..core import qaqc
 
 
 def cdf_to_nc(cdf_filename, atmpres=None, writefile=True, format="NETCDF4"):
@@ -63,10 +63,10 @@ def cdf_to_nc(cdf_filename, atmpres=None, writefile=True, format="NETCDF4"):
         ds = trim_min(ds, v)
 
     if "Turb" in ds:
-        ds = exo.trim_min(ds, "Turb")
-        ds = exo.trim_max(ds, "Turb")
-        ds = exo.trim_min_diff(ds, "Turb")
-        ds = exo.trim_max_diff(ds, "Turb")
+        ds = qaqc.trim_min(ds, "Turb")
+        ds = qaqc.trim_max(ds, "Turb")
+        ds = qaqc.trim_min_diff(ds, "Turb")
+        ds = qaqc.trim_max_diff(ds, "Turb")
 
     ds = utils.add_min_max(ds)
